Create Filtering a column-oriented database
Filtering a column-oriented database
Similar to before, you'll query data from the olympic_medals table using Snowflake, this time, filtering the table to only show a subset of rows. A connection object has been created, which is stored in the variable conn. Good luck!

Instructions

SELECT the team, name, and year columns (in that order) from the olympic_medals table.
Filter the dataset to only return rows after the year 2000.
Execute the query stored in the query variable, and print the results.