Create Querying nested semi-structured data
Within columns of type VARIANT, data can often be nested, as we see in the city_meta column of the host_cities table below.

Snowflake table with a single column of type VARIANT.

A Snowflake connection object to the database olympics has been created, and is available in the variable conn. Happy querying!

Instructions

Using dot notation, complete the query to extract data from the nested lat field in the coordinates object.
Query the nested long field from the coordinates object in the city_meta column.
Execute the query, and print the results.