Create Executing queries with sqlalchemy and pandas
Executing queries with sqlalchemy and pandas
To work with data stored in Postgres tables, you'll use sqlalchemy's create_engine(), and pandas's read_sql() functions. To get the hang of these tools, you'll practice connecting to a Postgres database, and executing a query. Good luck!

Instructions

Update the connection URI to create a connection to the disneyland database, over port 5432.
Use pandas to read the results of the provided SQL query into a DataFrame, using the connection object created in the previous step.