Create Querying JSON and JSONB data from Postgres
Querying JSON and JSONB data from Postgres
With Postgres' built-in JSON and JSONB data types, it's easy to store and interact with semi-structured data in a Postgres table. In this exercise, you'll observe some of the tooling that Postgres offers to query data of type JSON from the nested_reviews table. Best of luck!

Instructions

Create a connection to the disneyland database with user repl, using sqlalchemy.
Execute the query stored in the query variable, using the previously-defined db_engine.
Output the review column of the results DataFrame, and observe the data that was returned.